requirements.txt
PySide6>=6.5
mss>=9.0
pillow>=10.0
pytesseract>=0.3.10
tencentcloud-sdk-python>=3.0.0
requests>=2.31
python-dotenv>=1.0
pynput>=1.7
